shop/requirements.txt
Django==3.2
django-split-settings==1.2.0
gunicorn==20.1.0
djangorestframework==3.12.4
python-dotenv==0.21.0
psycopg2-binary==2.9.3
Pillow==8.3.2
django-filter==21.1